Fix critical, high, and medium cppcheck issues
- Fixed potential buffer overflows in fscanf usage in `src/test/bloom.c` and `src/test/hashtable.c`.
- Fixed multiple memory leaks in `src/content/handlers/html/redraw.c` and `src/content/handlers/image/svg.c` by replacing `alloca` with `malloc` / `free`.
- Fixed memory leak in `src/content/handlers/image/nssprite.c`.
- Fixed null pointer dereferences and redundant checks in `src/content/handlers/html/layout.c`, `src/desktop/treeview.c`, `src/content/urldb.c`, and `src/utils/talloc.c`.
- Fixed incorrect printf format specifiers in `src/content/fetchers/about/chart.c`, `src/content/fs_backing_store.c`, `src/content/handlers/html/box_inspect.c`, `src/content/handlers/image/image_cache.c`, and `src/content/urldb.c`.
- Added suppression comments for false positive syntax errors in conditionally compiled blocks in `src/content/fetchers/about/testament.c`.
- Added suppression comments for talloc false positives in `src/utils/talloc.c`.

+++ src/content/handlers/image/image_cache.c
@@ -739,7 +739,7 @@ int image_cache_snsummaryf(char *string, size_t size, const char *fmt)
 
 
             case 'j':
-                slen += snprintf(string + slen, size - slen, "%u", pct ? 100 : op_count);
+                slen += snprintf(string + slen, size - slen, "%u", (unsigned int)(pct ? 100 : op_count));
                 break;
 
                 FMTPCHR('k', "d", hit_count, op_count);
